Extract out a variants test
