+++ crates/test-rust-wasm/src/bin/variants.rs
@@ -1,0 +1,3 @@
+include!("../../../../tests/runtime/variants/wasm.rs");
+
+fn main() {}
