+++ tests/runtime/imports.c
@@ -7,185 +7,6 @@
 #include <string.h>
 #include <wasm.h>
 
-static void test_integers() {
-
-  uint8_t a;
-  uint16_t b;
-  host_multiple_results(&a, &b);
-  assert(a == 4);
-  assert(b == 5);
-}
-
-static void test_floats() {
-}
-
-static void test_chars() {
-}
-
-static void test_get_set() {
-}
-
-
-static void test_variants() {
-  {
-    host_option_f32_t a;
-    uint8_t r;
-    a.tag = 1;
-    a.val = 1;
-    assert(host_roundtrip_option(&a, &r) && r == 1);
-    assert(r == 1);
-    a.tag = 0;
-    assert(!host_roundtrip_option(&a, &r));
-    a.tag = 2;
-    a.val = 2;
-    assert(host_roundtrip_option(&a, &r) && r == 2);
-  }
-
-
-  {
-    host_expected_u32_f32_t a;
-    host_expected_f64_u8_t b;
-
-    a.tag = 0;
-    a.val.ok = 2;
-    host_roundtrip_result(&a, &b);
-    assert(b.tag == 0);
-    assert(b.val.ok == 2.0);
-
-    a.val.ok = 4;
-    host_roundtrip_result(&a, &b);
-    assert(b.tag == 0);
-    assert(b.val.ok == 4);
-
-    a.tag = 1;
-    a.val.err = 5.3;
-    host_roundtrip_result(&a, &b);
-    assert(b.tag == 1);
-    assert(b.val.err == 5);
-  }
-
-  assert(host_roundtrip_enum(HOST_E1_A) == HOST_E1_A);
-  assert(host_roundtrip_enum(HOST_E1_B) == HOST_E1_B);
-
-  assert(host_invert_bool(true) == false);
-  assert(host_invert_bool(false) == true);
-
-  {
-    host_casts_t c;
-    host_c1_t r1;
-    host_c2_t r2;
-    host_c3_t r3;
-    host_c4_t r4;
-    host_c5_t r5;
-    host_c6_t r6;
-    c.f0.tag = HOST_C1_A;
-    c.f0.val.a = 1;
-    c.f1.tag = HOST_C2_A;
-    c.f1.val.a = 2;
-    c.f2.tag = HOST_C3_A;
-    c.f2.val.a = 3;
-    c.f3.tag = HOST_C4_A;
-    c.f3.val.a = 4;
-    c.f4.tag = HOST_C5_A;
-    c.f4.val.a = 5;
-    c.f5.tag = HOST_C6_A;
-    c.f5.val.a = 6;
-    host_variant_casts(&c, &r1, &r2, &r3, &r4, &r5, &r6);
-    assert(r1.tag == HOST_C1_A && r1.val.a == 1);
-    assert(r2.tag == HOST_C2_A && r2.val.a == 2);
-    assert(r3.tag == HOST_C3_A && r3.val.a == 3);
-    assert(r4.tag == HOST_C4_A && r4.val.a == 4);
-    assert(r5.tag == HOST_C5_A && r5.val.a == 5);
-    assert(r6.tag == HOST_C6_A && r6.val.a == 6);
-  }
-
-  {
-    host_casts_t c;
-    host_c1_t r1;
-    host_c2_t r2;
-    host_c3_t r3;
-    host_c4_t r4;
-    host_c5_t r5;
-    host_c6_t r6;
-    c.f0.tag = HOST_C1_B;
-    c.f0.val.b = 1;
-    c.f1.tag = HOST_C2_B;
-    c.f1.val.b = 2;
-    c.f2.tag = HOST_C3_B;
-    c.f2.val.b = 3;
-    c.f3.tag = HOST_C4_B;
-    c.f3.val.b = 4;
-    c.f4.tag = HOST_C5_B;
-    c.f4.val.b = 5;
-    c.f5.tag = HOST_C6_B;
-    c.f5.val.b = 6;
-    host_variant_casts(&c, &r1, &r2, &r3, &r4, &r5, &r6);
-    assert(r1.tag == HOST_C1_B && r1.val.b == 1);
-    assert(r2.tag == HOST_C2_B && r2.val.b == 2);
-    assert(r3.tag == HOST_C3_B && r3.val.b == 3);
-    assert(r4.tag == HOST_C4_B && r4.val.b == 4);
-    assert(r5.tag == HOST_C5_B && r5.val.b == 5);
-    assert(r6.tag == HOST_C6_B && r6.val.b == 6);
-  }
-
-  {
-    host_zeros_t c;
-    host_z1_t r1;
-    host_z2_t r2;
-    host_z3_t r3;
-    host_z4_t r4;
-    c.f0.tag = HOST_Z1_A;
-    c.f0.val.a = 1;
-    c.f1.tag = HOST_Z2_A;
-    c.f1.val.a = 2;
-    c.f2.tag = HOST_Z3_A;
-    c.f2.val.a = 3;
-    c.f3.tag = HOST_Z4_A;
-    c.f3.val.a = 4;
-    host_variant_zeros(&c, &r1, &r2, &r3, &r4);
-    assert(r1.tag == HOST_Z1_A && r1.val.a == 1);
-    assert(r2.tag == HOST_Z2_A && r2.val.a == 2);
-    assert(r3.tag == HOST_Z3_A && r3.val.a == 3);
-    assert(r4.tag == HOST_Z4_A && r4.val.a == 4);
-  }
-
-  {
-    host_zeros_t c;
-    host_z1_t r1;
-    host_z2_t r2;
-    host_z3_t r3;
-    host_z4_t r4;
-    c.f0.tag = HOST_Z1_B;
-    c.f1.tag = HOST_Z2_B;
-    c.f2.tag = HOST_Z3_B;
-    c.f3.tag = HOST_Z4_B;
-    host_variant_zeros(&c, &r1, &r2, &r3, &r4);
-    assert(r1.tag == HOST_Z1_B);
-    assert(r2.tag == HOST_Z2_B);
-    assert(r3.tag == HOST_Z3_B);
-    assert(r4.tag == HOST_Z4_B);
-  }
-
-  {
-    host_option_typedef_t a;
-    a.tag = 0;
-    bool b = false;
-    host_result_typedef_t c;
-    c.tag = 1;
-    host_variant_typedefs(&a, b, &c);
-  }
-
-  {
-    bool a;
-    host_expected_void_void_t b;
-    host_my_errno_t c;
-    host_variant_enums(true, 0, HOST_MY_ERRNO_SUCCESS, &a, &b, &c);
-    assert(a == false);
-    assert(b == 1);
-    assert(c == HOST_MY_ERRNO_A);
-  }
-}
-
 static void test_lists() {
   {
     uint8_t list[] = {1, 2, 3, 4};
