+++ tests/runtime/wasmtime-run/exports.rs
@@ -26,68 +26,6 @@ pub fn test(wasm: &Wasm<Context>, instance: Instance, store: &mut Store<Context>
     Ok(())
 }
 
-fn scalars(wasm: &Wasm<Context>, store: &mut Store<Context>) -> Result<()> {
-    Ok(())
-}
-
-fn records(wasm: &Wasm<Context>, store: &mut Store<Context>) -> Result<()> {
-    Ok(())
-}
-
-fn variants(wasm: &Wasm<Context>, store: &mut Store<Context>) -> Result<()> {
-    assert_eq!(wasm.roundtrip_option(&mut *store, Some(1.0))?, Some(1));
-    assert_eq!(wasm.roundtrip_option(&mut *store, None)?, None);
-    assert_eq!(wasm.roundtrip_option(&mut *store, Some(2.0))?, Some(2));
-    assert_eq!(wasm.roundtrip_result(&mut *store, Ok(2))?, Ok(2.0));
-    assert_eq!(wasm.roundtrip_result(&mut *store, Ok(4))?, Ok(4.0));
-    assert_eq!(wasm.roundtrip_result(&mut *store, Err(5.3))?, Err(5));
-
-    assert_eq!(wasm.roundtrip_enum(&mut *store, E1::A)?, E1::A);
-    assert_eq!(wasm.roundtrip_enum(&mut *store, E1::B)?, E1::B);
-
-    assert_eq!(wasm.invert_bool(&mut *store, true)?, false);
-    assert_eq!(wasm.invert_bool(&mut *store, false)?, true);
-
-    let (a1, a2, a3, a4, a5, a6) = wasm.variant_casts(
-        &mut *store,
-        (C1::A(1), C2::A(2), C3::A(3), C4::A(4), C5::A(5), C6::A(6.0)),
-    )?;
-    assert!(matches!(a1, C1::A(1)));
-    assert!(matches!(a2, C2::A(2)));
-    assert!(matches!(a3, C3::A(3)));
-    assert!(matches!(a4, C4::A(4)));
-    assert!(matches!(a5, C5::A(5)));
-    assert!(matches!(a6, C6::A(b) if b == 6.0));
-
-    let (a1, a2, a3, a4, a5, a6) = wasm.variant_casts(
-        &mut *store,
-        (
-            C1::B(1),
-            C2::B(2.0),
-            C3::B(3.0),
-            C4::B(4.0),
-            C5::B(5.0),
-            C6::B(6.0),
-        ),
-    )?;
-    assert!(matches!(a1, C1::B(1)));
-    assert!(matches!(a2, C2::B(b) if b == 2.0));
-    assert!(matches!(a3, C3::B(b) if b == 3.0));
-    assert!(matches!(a4, C4::B(b) if b == 4.0));
-    assert!(matches!(a5, C5::B(b) if b == 5.0));
-    assert!(matches!(a6, C6::B(b) if b == 6.0));
-
-    let (a1, a2, a3, a4) =
-        wasm.variant_zeros(&mut *store, (Z1::A(1), Z2::A(2), Z3::A(3.0), Z4::A(4.0)))?;
-    assert!(matches!(a1, Z1::A(1)));
-    assert!(matches!(a2, Z2::A(2)));
-    assert!(matches!(a3, Z3::A(b) if b == 3.0));
-    assert!(matches!(a4, Z4::A(b) if b == 4.0));
-
-    wasm.variant_typedefs(&mut *store, None, false, Err(()))?;
-    Ok(())
-}
-
 fn lists(wasm: &Wasm<Context>, store: &mut Store<Context>) -> Result<()> {
     wasm.list_param(&mut *store, &[1, 2, 3, 4])?;
     wasm.list_param2(&mut *store, "foo")?;
